rag-bot/requirements.txt
fastapi
uvicorn[standard]
streamlit
requests
python-dotenv
langchain
langchain-openai
langchain-core
langchain-community
langchain-text-splitters
faiss-cpu
tiktoken
azure-identity
